requirements.txt
rich
pyyaml
discord
cogwatch
pymongo[srv]
python-dotenv